+++ R/overview.R
@@ -8,51 +8,64 @@
 #'
 #' @returns
 #' `overview(x)` returns a 1-row data frame unless
-#' `x` is a data frame, in which a list of data frames is returned.
+#' `x` is a data frame, in which case an object of class "overview" is returned,
+#' Under the hood this is just a a list of data frames.
 #' Key summary statistics are reported in each data frame.
 #'
+#' @rdname overview
 #' @export
 overview <- function(x, hist = FALSE){
  UseMethod("overview")
 }
+#' @rdname overview
 #' @export
 overview.default <- function(x, hist = FALSE){
-  warning("Unsure how to calculate summary for x, falling back to character")
-  out <- as.character(x)
-  out <- overview(list_as_df(list(x = out)), hist = hist)$categorical
+  out <- overview(list_as_df(list(x = x)), hist = hist)$other
   names(out)[1] <- "length"
   out[[1]] <- length(x)
-  out[[2]] <- utils::tail(class(x), n = 1)
   out
 }
+#' @rdname overview
 #' @export
 overview.logical <- function(x, hist = FALSE){
-  out <- overview(list_as_df(list(x = x)), hist = hist)$logical
+  out <- overview(list_as_df(list(x = as.logical(x))), hist = hist)$logical
   names(out)[1] <- "length"
   out[[1]] <- length(x)
   out
 }
+#' @rdname overview
 #' @export
 overview.numeric <- function(x, hist = FALSE){
-  out <- overview(list_as_df(list(x = x)), hist = hist)$numeric
+  out <- overview(list_as_df(list(x = as.numeric(x))), hist = hist)$numeric
   names(out)[1] <- "length"
   out[[1]] <- length(x)
   out
 }
+#' @rdname overview
 #' @export
 overview.character <- function(x, hist = FALSE){
-  out <- overview(list_as_df(list(x = x)), hist = hist)$categorical
+  out <- overview(list_as_df(list(x = as.character(x))), hist = hist)$categorical
   names(out)[1] <- "length"
   out[[1]] <- length(x)
   out
 }
+#' @rdname overview
+#' @export
+overview.factor <- function(x, hist = FALSE){
+  out <- overview(list_as_df(list(x = as.factor(x))), hist = hist)$categorical
+  names(out)[1] <- "length"
+  out[[1]] <- length(x)
+  out
+}
+#' @rdname overview
 #' @export
 overview.Date <- function(x, hist = FALSE){
-  out <- overview(list_as_df(list(x = x)), hist = hist)$date
+  out <- overview(list_as_df(list(x = as.Date(x))), hist = hist)$date
   names(out)[1] <- "length"
   out[[1]] <- length(x)
   out
 }
+#' @rdname overview
 #' @export
 overview.POSIXt <- function(x, hist = FALSE){
   out <- overview(list_as_df(list(x = as.POSIXct(x))), hist = hist)$datetime
@@ -61,15 +74,16 @@ overview.POSIXt <- function(x, hist = FALSE){
   out[[2]] <- utils::tail(class(x), n = 1)
   out
 }
+#' @rdname overview
 #' @export
 overview.data.frame <- function(x, hist = FALSE){
   check_is_df(x)
   N <- nrow(x)
   num_cols <- ncol(x)
-  skim_df <- df_as_tbl(list_as_df(x))
+  skim_df <- x
   data_nms <- names(skim_df)
   col_classes <- vapply(skim_df, function(x) utils::tail(class(x), n = 1), "")
-  out <- df_as_tbl(enframe_(col_classes, name = "col", value = "class"))
+  out <- list_as_df(enframe_(col_classes, name = "col", value = "class"))
   chr_vars <- data_nms[vapply(skim_df, is.character, FALSE,
                               USE.NAMES = FALSE)]
   if (length(chr_vars) > 0L) {
@@ -83,27 +97,29 @@ overview.data.frame <- function(x, hist = FALSE){
   date_vars <- data_nms[vapply(skim_df, function(x) inherits(x, "Date"), FALSE)]
   datetime_vars <- data_nms[vapply(skim_df, function(x) inherits(x, "POSIXt"),  FALSE)]
   cat_vars <- data_nms[vapply(skim_df, is.factor, FALSE)]
-  other_vars <- setdiff(data_nms, c(lgl_vars, num_vars, date_vars,
-                                    datetime_vars, cat_vars))
-  if (length(other_vars) > 0) {
-    warning(paste0("Unsure how to calculate summaries for these variables: \n",
-                   paste(other_vars, collapse = "\n")), "\n\nFalling back to character")
-    for (var in other_vars) {
-      skim_df[[var]] <- factor_(as.character(skim_df[[var]]), ordered = TRUE)
-    }
-    cat_vars <- c(cat_vars, other_vars)
-    cat_vars <- data_nms[sort(match(cat_vars, data_nms))]
-  }
+  other_vars <- setdiff_(data_nms, c(lgl_vars, num_vars, date_vars,
+                                     datetime_vars, cat_vars))
+  # if (length(other_vars) > 0) {
+  #   warning(paste0("Unsure how to calculate summaries for these variables: \n",
+  #                  paste(other_vars, collapse = "\n")), "\n\nFalling back to character")
+  #   for (var in other_vars) {
+  #     skim_df[[var]] <- factor_(as.character(skim_df[[var]]), ordered = TRUE)
+  #   }
+  #   cat_vars <- c(cat_vars, other_vars)
+  #   cat_vars <- data_nms[sort(match(cat_vars, data_nms))]
+  # }
 
 # Logical -----------------------------------------------------------------
 
   lgl_data <- df_select(skim_df, lgl_vars)
   which_lgl <- which_in(out[["col"]], lgl_vars)
   lgl_out <- out[which_lgl, , drop = FALSE]
-  lgl_out <- df_add_cols(lgl_out, list(n_missing = NA_integer_))
-  lgl_out <- df_add_cols(lgl_out, list(p_complete = NA_real_))
-  lgl_out <- df_add_cols(lgl_out, list(n_true = NA_integer_, n_false = NA_integer_))
-  lgl_out <- df_add_cols(lgl_out, list(p_true = NA_real_))
+  value_size <- min(length(which_lgl), 1L)
+  lgl_out <- df_add_cols(lgl_out, list(n_missing = NA_integer_[value_size]))
+  lgl_out <- df_add_cols(lgl_out, list(p_complete = NA_real_[value_size]))
+  lgl_out <- df_add_cols(lgl_out, list(n_true = NA_integer_[value_size],
+                                       n_false = NA_integer_[value_size]))
+  lgl_out <- df_add_cols(lgl_out, list(p_true = NA_real_[value_size]))
   if (N > 0L && length(which_lgl) > 0) {
     lgl_out$n_missing <- pluck_row(summarise_all(lgl_data, num_na), 1)
     lgl_out$p_complete <- pluck_row(summarise_all(lgl_data, prop_complete), 1)
@@ -111,21 +127,22 @@ overview.data.frame <- function(x, hist = FALSE){
     lgl_out$n_false <- N - lgl_out[["n_missing"]] - lgl_out[["n_true"]]
     lgl_out$p_true <- lgl_out[["n_true"]] / (N - lgl_out[["n_missing"]])
   }
-  lgl_out <- df_as_tbl(lgl_out)
 
 # Numeric -----------------------------------------------------------------
 
   num_data <- df_select(skim_df, num_vars)
   which_num <- which_in(out[["col"]], num_vars)
   num_out <- out[which_num, , drop = FALSE]
-  num_out <- df_add_cols(num_out, list(n_missing = NA_integer_))
-  num_out <- df_add_cols(num_out, list(p_complete = NA_real_))
-  num_out <- df_add_cols(num_out, list(n_unique = NA_integer_))
-  num_out <- df_add_cols(num_out, stats::setNames(as.list(rep_len(NA_real_,
-                                                  8)), c("mean", "p0", "p25", "p50", "p75", "p100", "iqr",
-                                                         "sd")))
+  value_size <- min(length(which_num), 1L)
+  num_out <- df_add_cols(num_out, list(n_missing = NA_integer_[value_size]))
+  num_out <- df_add_cols(num_out, list(p_complete = NA_real_[value_size]))
+  num_out <- df_add_cols(num_out, list(n_unique = NA_integer_[value_size]))
+  num_out <- df_add_cols(num_out, stats::setNames(
+    new_list(8, default = NA_real_[value_size]),
+    c("mean", "p0", "p25", "p50", "p75", "p100", "iqr", "sd")
+  ))
   if (hist){
-    num_out$hist <- NA_character_
+    num_out$hist <- NA_character_[value_size]
   }
   if (N > 0L && length(which_num) > 0) {
     num_out$n_missing <- pluck_row(summarise_all(num_data, num_na), 1)
@@ -152,18 +169,18 @@ overview.data.frame <- function(x, hist = FALSE){
       ), 1)
     }
   }
-  num_out <- df_as_tbl(num_out)
 
 # Dates -------------------------------------------------------------------
 
   date_data <- df_select(skim_df, date_vars)
   which_date <- which_in(out[["col"]], date_vars)
   date_out <- out[which_date, , drop = FALSE]
-  date_out <- df_add_cols(date_out, list(n_missing = NA_integer_))
-  date_out <- df_add_cols(date_out, list(p_complete = NA_real_))
-  date_out <- df_add_cols(date_out, list(n_unique = NA_integer_))
-  date_out <- df_add_cols(date_out, list(min = .Date(NA_real_),
-                                         max = .Date(NA_real_)))
+  value_size <- min(length(which_date), 1L)
+  date_out <- df_add_cols(date_out, list(n_missing = NA_integer_[value_size]))
+  date_out <- df_add_cols(date_out, list(p_complete = NA_real_[value_size]))
+  date_out <- df_add_cols(date_out, list(n_unique = NA_integer_[value_size]))
+  date_out <- df_add_cols(date_out, list(min = .Date(NA_real_[value_size]),
+                                         max = .Date(NA_real_[value_size])))
   if (N > 0L && length(which_date) > 0) {
     date_out$n_missing <- pluck_row(summarise_all(date_data, num_na), 1)
     date_out$p_complete <- pluck_row(summarise_all(date_data, prop_complete), 1)
@@ -174,7 +191,6 @@ overview.data.frame <- function(x, hist = FALSE){
     class(date_out$min) <- "Date"
     class(date_out$max) <- "Date"
   }
-  date_out <- df_as_tbl(date_out)
 
 
 # Date-Times --------------------------------------------------------------
@@ -183,11 +199,13 @@ overview.data.frame <- function(x, hist = FALSE){
   datetime_data <- transform_all(datetime_data, as.POSIXct)
   which_datetime <- which_in(out[["col"]], datetime_vars)
   datetime_out <- out[which_datetime, , drop = FALSE]
-  datetime_out <- df_add_cols(datetime_out, list(n_missing = NA_integer_))
-  datetime_out <- df_add_cols(datetime_out, list(p_complete = NA_real_))
-  datetime_out <- df_add_cols(datetime_out, list(n_unique = NA_integer_))
-  datetime_out <- df_add_cols(datetime_out, list(min = .POSIXct(NA_real_),
-                                                 max = .POSIXct(NA_real_)))
+  value_size <- min(length(which_datetime), 1L)
+  datetime_out <- df_add_cols(datetime_out, list(n_missing = NA_integer_[value_size]))
+  datetime_out <- df_add_cols(datetime_out, list(p_complete = NA_real_[value_size]))
+  datetime_out <- df_add_cols(datetime_out, list(n_unique = NA_integer_[value_size]))
+  datetime_out <- df_add_cols(datetime_out, list(tzone = NA_character_[value_size]))
+  datetime_out <- df_add_cols(datetime_out, list(min = .POSIXct(NA_real_[value_size]),
+                                                 max = .POSIXct(NA_real_[value_size])))
   if (N > 0L && length(which_datetime) > 0) {
     datetime_out$n_missing <- pluck_row(summarise_all(datetime_data, num_na), 1)
     datetime_out$p_complete <- pluck_row(summarise_all(datetime_data, prop_complete), 1)
@@ -197,8 +215,10 @@ overview.data.frame <- function(x, hist = FALSE){
     datetime_out$max <- pluck_row(summarise_all(datetime_data, collapse::fmax), 1)
     datetime_out$min <- .POSIXct(datetime_out$min, tz = "UTC")
     datetime_out$max <- .POSIXct(datetime_out$max, tz = "UTC")
+    for (i in seq_len(nrow(datetime_out))){
+      datetime_out[["tzone"]][i] <- tzone(x[[datetime_out[["col"]][i]]])
+    }
   }
-  datetime_out <- df_as_tbl(datetime_out)
 
 
 # Categorical -------------------------------------------------------------
@@ -206,11 +226,13 @@ overview.data.frame <- function(x, hist = FALSE){
   cat_data <- df_select(skim_df, cat_vars)
   which_cat <- which_in(out[["col"]], cat_vars)
   cat_out <- out[which_cat, , drop = FALSE]
-  cat_out <- df_add_cols(cat_out, list(n_missing = NA_integer_))
-  cat_out <- df_add_cols(cat_out, list(p_complete = NA_real_))
-  cat_out <- df_add_cols(cat_out, list(n_unique = NA_integer_))
-  cat_out <- df_add_cols(cat_out, stats::setNames(as.list(rep_len(NA_character_, 2)),
-                                            c("min", "max")))
+  value_size <- min(length(which_cat), 1L)
+  cat_out <- df_add_cols(cat_out, list(n_missing = NA_integer_[value_size]))
+  cat_out <- df_add_cols(cat_out, list(p_complete = NA_real_[value_size]))
+  cat_out <- df_add_cols(cat_out, list(n_unique = NA_integer_[value_size]))
+  cat_out <- df_add_cols(cat_out, list(n_levels = NA_integer_[value_size]))
+  cat_out <- df_add_cols(cat_out, list(min = NA_character_[value_size],
+                                       max = NA_character_[value_size]))
   if (N > 0L && length(which_cat) > 0) {
     cat_out$n_missing <- pluck_row(summarise_all(cat_data, num_na), 1)
     cat_out$p_complete <- pluck_row(summarise_all(cat_data, prop_complete), 1)
@@ -218,25 +240,123 @@ overview.data.frame <- function(x, hist = FALSE){
     cat_out$n_unique <- cat_out$n_unique - (cat_out$n_missing > 0L)
     cat_out$min <- pluck_row(summarise_all(cat_data, collapse::fmin), 1)
     cat_out$max <- pluck_row(summarise_all(cat_data, collapse::fmax), 1)
+    cat_out$min <- as.character(cat_out$min)
+    cat_out$max <- as.character(cat_out$max)
+    for (i in seq_len(nrow(cat_out))){
+      if (cat_out[["class"]][i] == "factor"){
+        cat_out[["n_levels"]][i] <- length(levels(x[[cat_out[["col"]][i]]]))
+      }
+    }
   }
-  cat_out <- df_as_tbl(cat_out)
-  list(
+
+  # Other -------------------------------------------------------------
+
+  other_data <- df_select(skim_df, other_vars)
+  which_other <- which_in(out[["col"]], other_vars)
+  other_out <- out[which_other, , drop = FALSE]
+  value_size <- min(length(which_other), 1L)
+  other_out <- df_add_cols(other_out, list(n_missing = NA_integer_[value_size]))
+  other_out <- df_add_cols(other_out, list(p_complete = NA_real_[value_size]))
+  other_out <- df_add_cols(other_out, list(n_unique = NA_integer_[value_size]))
+  if (N > 0L && length(which_other) > 0) {
+    other_out$n_missing <- pluck_row(summarise_all(
+      other_data, function(x) num_na(x, recursive = FALSE)
+      ), 1)
+    other_out$p_complete <- pluck_row(summarise_all(
+      other_data, function(x) prop_complete(x, recursive = FALSE)
+      ), 1)
+    other_out$n_unique <- pluck_row(summarise_all(
+      other_data, function(x) length(unique(x))
+      ), 1)
+    other_out$n_unique <- other_out$n_unique - (other_out$n_missing > 0L)
+  }
+
+  out <- list(
     nrow = N, ncol = num_cols,
     logical = lgl_out,
     numeric = num_out,
     date = date_out,
     datetime = datetime_out,
-    categorical = cat_out
+    categorical = cat_out,
+    other = other_out
   )
+  class(out) <- c("overview", "list")
+  out
 }
-
+#' @export
+print.overview <- function(x, max = NULL, ...){
+  # max_rows <- getOption("tibble.print_max", 20)
+  # max_cols <- getOption("tibble.width", NULL)
+  # max_extra_cols <- getOption("tibble.max_extra_cols", 100)
+  # options(tibble.print_max = 10)
+  # options(tibble.width = 100)
+  # options(tibble.max_extra_cols = 10)
+  cat(paste("rows:", x$nrow, "cols:", x$ncol), "\n")
+  # for (data_type in names(x)[-(1:2)]){
+  #   if (nrow(x[[data_type]])){
+  #     cat(paste("\n-----", data_type, "-----\n"))
+  #     print(x[[data_type]])
+  #   }
+  # }
+  if (nrow(x$logical)){
+    x$logical$p_complete <- pretty_num(round(x$logical$p_complete, 2))
+    cat("\n----- Logical -----\n")
+    print(x$logical)
+  }
+  if (nrow(x$numeric)){
+    x$numeric$p_complete <- pretty_num(round(x$numeric$p_complete, 2))
+    x$numeric$mean <- pretty_num(round(x$numeric$mean, 2))
+    x$numeric$p0 <- pretty_num(round(x$numeric$p0, 2))
+    x$numeric$p25 <- pretty_num(round(x$numeric$p25, 2))
+    x$numeric$p50 <- pretty_num(round(x$numeric$p50, 2))
+    x$numeric$p75 <- pretty_num(round(x$numeric$p75, 2))
+    x$numeric$p100 <- pretty_num(round(x$numeric$p100, 2))
+    x$numeric$iqr <- pretty_num(round(x$numeric$iqr, 2))
+    x$numeric$sd <- pretty_num(round(x$numeric$sd, 2))
+    cat("\n----- Numeric -----\n")
+    print(x$numeric)
+  }
+  if (nrow(x$date)){
+    x$date$p_complete <- pretty_num(round(x$date$p_complete, 2))
+    cat("\n----- Dates -----\n")
+    print(x$date)
+  }
+  if (nrow(x$datetime)){
+    x$datetime$p_complete <- pretty_num(round(x$datetime$p_complete, 2))
+    cat("\n----- Date-times -----\n")
+    print(x$datetime)
+  }
+  if (nrow(x$categorical)){
+    x$categorical$p_complete <- pretty_num(round(x$categorical$p_complete, 2))
+    cat("\n----- Categorical -----\n")
+    print(x$categorical)
+  }
+  if (nrow(x$other)){
+    x$other$p_complete <- pretty_num(round(x$other$p_complete, 2))
+    cat("\n----- Other -----\n")
+    print(x$other)
+  }
+  # options(tibble.print_max = max_rows)
+  # options(tibble.width = max_cols)
+  # options(tibble.max_extra_cols = max_extra_cols)
+  invisible(x)
+}
 ### Helpers
-n_unique <- function(x){
-  collapse::fnunique(x)
+
+n_unique <- function(x, na_rm = FALSE){
+  out <- collapse::fnunique(x)
+  if (na_rm){
+    out <- out - any_na(x, recursive = FALSE)
+  }
+  out
 }
-prop_complete <- function(x){
-  N <- unlisted_length(x)
-  1 - (num_na(x) / N)
+prop_complete <- function(x, recursive = TRUE){
+  if (recursive){
+    N <- unlisted_length(x)
+  } else {
+    N <- length(x)
+  }
+  1 - (num_na(x, recursive = recursive) / N)
 }
 transform_all <- function(data, .fn){
   for (col in names(data)){
@@ -246,6 +366,7 @@ transform_all <- function(data, .fn){
 }
 summarise_all <- function(data, .fn, size = 1){
   out <- data[seq_len(size), , drop = FALSE]
+  attr(out, "row.names") <- .set_row_names(size)
   for (col in names(out)){
     out[[col]] <- .fn(data[[col]])
   }
@@ -268,6 +389,7 @@ spark_bar <- function(x){
   out <- bars[bar_codes]
   paste0(out, collapse = "")
 }
+
 inline_hist <- function(x, n_bins = 5L){
   if (length(x) < 1L) {
     return(" ")
@@ -285,4 +407,7 @@ inline_hist <- function(x, n_bins = 5L){
                       nbins = n_bins)
   hist_dt <- hist_dt / max(hist_dt)
   spark_bar(hist_dt)
+}
+pretty_num <- function(x, scientific = FALSE, ...){
+  prettyNum(x, scientific = scientific, ...)
 }